pyproject.toml
[project]
name = "report-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "ipython>=9.4.0",
    "langchain>=0.1.17",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.2",
    "markdown2>=2.5.4",
    "pdfkit>=1.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
]
